Rename files to avoid clashes with other libs
