+++ cpp/OPLogs.h
