+++ cpp/OPMacros.hpp
