+++ cpp/OPPreparedStatementHostObject.cpp
@@ -1,11 +1,11 @@
-#include "PreparedStatementHostObject.hpp"
+#include "OPPreparedStatementHostObject.hpp"
 #if OP_SQLITE_USE_LIBSQL
-#include "libsql/bridge.hpp"
+#include "libsql/OPLibsqlBridge.hpp"
 #else
-#include "bridge.hpp"
+#include "OPBridge.hpp"
 #endif
-#include "macros.hpp"
-#include "utils.hpp"
+#include "OPMacros.hpp"
+#include "OPUtils.hpp"
 
 namespace opsqlite {
