+++ cpp/OPPreparedStatementHostObject.hpp
@@ -5,7 +5,7 @@
 #include <memory>
 #ifdef OP_SQLITE_USE_LIBSQL
 #include "libsql.h"
-#include "libsql/bridge.hpp"
+#include "libsql/OPLibsqlBridge.hpp"
 #else
 #ifdef __ANDROID__
 #include "sqlite3.h"
