+++ cpp/OPSmartHostObject.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "types.hpp"
+#include "OPTypes.hpp"
 #include <any>
 #include <jsi/jsi.h>
 #include <vector>
