+++ cpp/OPSqlite.cpp
@@ -1,15 +1,15 @@
 #include "OPSqlite.hpp"
 #include "OPDatabase.hpp"
-#include "DumbHostObject.hpp"
+#include "OPDumbHostObject.hpp"
 #include "OPThreadPool.hpp"
 #ifdef OP_SQLITE_USE_LIBSQL
-#include "libsql/bridge.hpp"
+#include "libsql/OPLibsqlBridge.hpp"
 #else
-#include "bridge.hpp"
+#include "OPBridge.hpp"
 #endif
-#include "logs.h"
-#include "macros.hpp"
-#include "utils.hpp"
+#include "OPLogs.h"
+#include "OPMacros.hpp"
+#include "OPUtils.hpp"
 #include <functional>
 #include <iostream>
 #include <string>
