+++ cpp/libsql/OPLibsqlBridge.cpp
@@ -1,9 +1,9 @@
-#include "bridge.hpp"
-#include "DumbHostObject.hpp"
-#include "SmartHostObject.hpp"
+#include "OPBridge.hpp"
+#include "OPDumbHostObject.hpp"
+#include "OPSmartHostObject.hpp"
 #include "libsql.h"
-#include "logs.h"
-#include "utils.hpp"
+#include "OPLogs.h"
+#include "OPUtils.hpp"
 #include <filesystem>
 #include <iostream>
 #include <unordered_map>
