+++ op-sqlite.podspec
@@ -108,10 +108,10 @@ Pod::Spec.new do |s|
   source_files = Dir.glob("ios/**/*.{h,hpp,m,mm}") + Dir.glob("cpp/**/*.{hpp,h,cpp,c}")
 
   # Backend bridges are selected explicitly by flags and should not be compiled by default.
-  source_files.reject! { |path| path == "cpp/turso/turso_bridge.cpp" } unless use_turso
+  source_files.reject! { |path| path == "cpp/turso/OPTursoBridge.cpp" } unless use_turso
 
   # Strictly blocks all headers from being public
-  s.public_header_files = ["ios/OPSQLite.h", "cpp/sqlite3.h", "cpp/bridge.hpp"]
+  s.public_header_files = ["ios/OPSQLite.h", "cpp/sqlite3.h", "cpp/OPBridge.hpp"]
 
   # Set the path to the `c_sources` directory based on environment
   if is_user_app
@@ -144,18 +144,18 @@ Pod::Spec.new do |s|
 
   if use_sqlcipher then
     log_message.call("[OP-SQLITE] using SQLCipher")
-    exclude_files += ["cpp/sqlite3.c", "cpp/sqlite3.h", "cpp/libsql/bridge.cpp", "cpp/libsql/bridge.hpp", "ios/libsql.xcframework/**/*", "ios/libsql_experimental.xcframework/**/*", "ios/turso_sdk_kit.xcframework/**/*"]
+    exclude_files += ["cpp/sqlite3.c", "cpp/sqlite3.h", "cpp/libsql/OPLibsqlBridge.cpp", "cpp/libsql/OPLibsqlBridge.hpp", "ios/libsql.xcframework/**/*", "ios/libsql_experimental.xcframework/**/*", "ios/turso_sdk_kit.xcframework/**/*"]
     xcconfig[:GCC_PREPROCESSOR_DEFINITIONS] += " OP_SQLITE_USE_SQLCIPHER=1 HAVE_FULLFSYNC=1 SQLITE_HAS_CODEC SQLITE_TEMP_STORE=3 SQLITE_EXTRA_INIT=sqlcipher_extra_init SQLITE_EXTRA_SHUTDOWN=sqlcipher_extra_shutdown"
     s.dependency "OpenSSL-Universal"
   elsif use_turso then
     log_message.call("[OP-SQLITE] using Turso SDK kit")
-    exclude_files += ["cpp/sqlite3.c", "cpp/sqlite3.h", "cpp/bridge.hpp", "cpp/bridge.cpp", "cpp/sqlcipher/sqlite3.c", "cpp/sqlcipher/sqlite3.h", "cpp/libsql/bridge.hpp", "cpp/libsql/bridge.cpp", "ios/libsql_experimental.xcframework/**/*"]
+    exclude_files += ["cpp/sqlite3.c", "cpp/sqlite3.h", "cpp/OPBridge.hpp", "cpp/OPBridge.cpp", "cpp/sqlcipher/sqlite3.c", "cpp/sqlcipher/sqlite3.h", "cpp/libsql/OPLibsqlBridge.hpp", "cpp/libsql/OPLibsqlBridge.cpp", "ios/libsql_experimental.xcframework/**/*"]
   elsif use_libsql then
     log_message.call("[OP-SQLITE] ⚠️ Using libsql. If you have libsql questions please ask in the Turso Discord server.")
-    exclude_files += ["cpp/sqlite3.c", "cpp/sqlite3.h", "cpp/sqlcipher/sqlite3.c", "cpp/sqlcipher/sqlite3.h", "cpp/bridge.h", "cpp/bridge.cpp", "ios/turso_sdk_kit.xcframework/**/*"]
+    exclude_files += ["cpp/sqlite3.c", "cpp/sqlite3.h", "cpp/sqlcipher/sqlite3.c", "cpp/sqlcipher/sqlite3.h", "cpp/OPBridge.hpp", "cpp/OPBridge.cpp", "ios/turso_sdk_kit.xcframework/**/*"]
   else
     log_message.call("[OP-SQLITE] using pure SQLite")
-    exclude_files += ["cpp/sqlcipher/sqlite3.c", "cpp/sqlcipher/sqlite3.h", "cpp/libsql/bridge.c", "cpp/libsql/bridge.h", "cpp/libsql/bridge.cpp", "cpp/libsql/libsql.h", "ios/libsql_experimental.xcframework/**/*", "ios/turso_sdk_kit.xcframework/**/*"]
+    exclude_files += ["cpp/sqlcipher/sqlite3.c", "cpp/sqlcipher/sqlite3.h", "cpp/libsql/OPLibsqlBridge.hpp", "cpp/libsql/OPLibsqlBridge.cpp", "cpp/libsql/libsql.h", "ios/libsql_experimental.xcframework/**/*", "ios/turso_sdk_kit.xcframework/**/*"]
   end
 
    # Exclude xcframeworks that aren't being used
